requirements.txt
fastapi
uvicorn
opencv-python
python-multipart
pillow